i32 -> isize

+++ src/main.rs
@@ -21,8 +21,8 @@ struct Bitmap {
     pixels: Vec<bool>,
 }
 struct Drawable {
-    x: i32,
-    y: i32,
+    x: isize,
+    y: isize,
     bitmap: Bitmap,
 }
 impl Bitmap {
@@ -33,11 +33,11 @@ impl Bitmap {
             pixels: (0..(w as usize) * (h as usize)).map(|_| on).collect::<Vec<bool>>(),
         }
     }
-    fn get(&self, x: i32, y: i32) -> bool {
-        x >= 0 && y >= 0 && x < self.w as i32 && y < self.h as i32 && self.pixels[y as usize * self.w + x as usize]
+    fn get(&self, x: isize, y: isize) -> bool {
+        x >= 0 && y >= 0 && x < self.w as isize && y < self.h as isize && self.pixels[y as usize * self.w + x as usize]
     }
-    fn set(&mut self, x: i32, y: i32, on: bool) {
-        if x >= 0 && y >= 0 && x < self.w as i32 && y < self.h as i32 {
+    fn set(&mut self, x: isize, y: isize, on: bool) {
+        if x >= 0 && y >= 0 && x < self.w as isize && y < self.h as isize {
             self.pixels[y as usize * self.w + x as usize] = on;
         }
     }
@@ -113,16 +113,16 @@ impl Drawable {
         Drawable {
             x: match x {
                 DrawPos::Coord(p) => p,
-                DrawPos::Center => (SCREEN_WIDTH as i32 - bitmap.w as i32) / 2,
+                DrawPos::Center => (SCREEN_WIDTH as isize - bitmap.w as isize) / 2,
             },
             y: match y {
                 DrawPos::Coord(p) => p,
-                DrawPos::Center => (SCREEN_HEIGHT as i32 - bitmap.h as i32) / 2,
+                DrawPos::Center => (SCREEN_HEIGHT as isize - bitmap.h as isize) / 2,
             },
             bitmap,
         }
     }
-    fn rect(x: i32, y: i32, w: usize, h: usize, on: bool) -> Drawable {
+    fn rect(x: isize, y: isize, w: usize, h: usize, on: bool) -> Drawable {
         Drawable {
             x,
             y,
@@ -130,8 +130,8 @@ impl Drawable {
         }
     }
     fn crop_to_screen(&self) -> Drawable {
-        let x = std::cmp::min(SCREEN_WIDTH - 1, std::cmp::max(0, self.x) as usize) as i32;
-        let y = std::cmp::min(SCREEN_HEIGHT - 1, std::cmp::max(0, self.y) as usize) as i32;
+        let x = std::cmp::min(SCREEN_WIDTH - 1, std::cmp::max(0, self.x) as usize) as isize;
+        let y = std::cmp::min(SCREEN_HEIGHT - 1, std::cmp::max(0, self.y) as usize) as isize;
         let w = std::cmp::min(SCREEN_WIDTH - x as usize, self.bitmap.w);
         let h = std::cmp::min(SCREEN_HEIGHT - y as usize, self.bitmap.h);
         let x_crop = (-std::cmp::min(0, self.x)) as usize;
@@ -146,9 +146,9 @@ impl Drawable {
         for bx in 0..other.bitmap.w {
             for by in 0..other.bitmap.h {
                 self.bitmap.set(
-                    other.x + bx as i32,
-                    other.y + by as i32,
-                    other.bitmap.get(bx as i32, by as i32),
+                    other.x + bx as isize,
+                    other.y + by as isize,
+                    other.bitmap.get(bx as isize, by as isize),
                 );
             }
         }
@@ -191,7 +191,7 @@ impl Drawable {
         let splits = self.bitmap.w.div_ceil(REPORT_SPLIT_SZ);
         for i in 0..splits {
             vec.push(Drawable {
-                x: self.x + (i * REPORT_SPLIT_SZ) as i32,
+                x: self.x + (i * REPORT_SPLIT_SZ) as isize,
                 y: self.y,
                 bitmap: self.bitmap.crop(
                     i * REPORT_SPLIT_SZ,
@@ -207,13 +207,13 @@ impl Drawable {
 
 #[derive(Clone, Copy)]
 enum DrawPos {
-    Coord(i32),
+    Coord(isize),
     Center,
 }
 impl FromStr for DrawPos {
     type Err = &'static str;
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        s.parse::<i32>().map(|n| Ok(DrawPos::Coord(n))).unwrap_or_else(|_| {
+        s.parse::<isize>().map(|n| Ok(DrawPos::Coord(n))).unwrap_or_else(|_| {
             if ["center", "c", "middle", "m"].contains(&s) {
                 Ok(DrawPos::Center)
             } else {
